backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
pdfplumber
camelot-py[cv]
pytesseract
opencv-python
faiss-cpu
sentence-transformers
transformers
torch
python-dotenv
gunicorn